+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/PluginsSerializer.kt
@@ -1,22 +1,40 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.utils
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.Plugin
+import com.github.insanusmokrassar.TelegramBotAPI.utils.toJson
 import kotlinx.serialization.*
 import kotlinx.serialization.internal.ArrayListSerializer
+import kotlinx.serialization.internal.StringDescriptor
+import kotlinx.serialization.json.*
+import kotlin.reflect.KClass
 
-object ListSerializer: KSerializer<List<Plugin>> {
-    private val realSerializer = ArrayListSerializer(
-        PolymorphicSerializer(Plugin::class)
-    )
+object PluginSerializer : KSerializer<Plugin> {
+    override val descriptor: SerialDescriptor = StringDescriptor.withName(Plugin::class.simpleName ?: "Plugin")
 
-    override val descriptor: SerialDescriptor
-        get() = realSerializer.descriptor
+    override fun serialize(encoder: Encoder, obj: Plugin) {
+        val array = JsonArray(
+            listOf(
+                JsonPrimitive(obj::class.java.canonicalName),
+                obj.toJson(serializerByTypeToken(obj::class.java))
+            )
+        )
+        JsonArraySerializer.serialize(
+            encoder,
+            array
+        )
+    }
 
-    override fun deserialize(decoder: Decoder): List<Plugin> = realSerializer.deserialize(
-        decoder
-    ).map { it as Plugin }
-
-    override fun serialize(encoder: Encoder, obj: List<Plugin>) = realSerializer.serialize(
-        encoder, obj
-    )
+    override fun deserialize(decoder: Decoder): Plugin {
+        val decoded = decoder.decode(JsonArraySerializer)
+        val expectedClass = decoded.getPrimitive(0).content
+        val expectedArgumentObject = decoded.getObjectOrNull(1)
+        val serializer = Class.forName(expectedClass).kotlin.serializer()
+        val argumentsObject = expectedArgumentObject ?.toString() ?: JsonObject(emptyMap()).toString()
+        return Json.nonstrict.parse(
+            serializer,
+            argumentsObject
+        ) as Plugin
+    }
 }
+
+object PluginsListSerializer: KSerializer<List<Plugin>> by ArrayListSerializer(PluginSerializer)